[수정] mainStore 로직과 filterStore 로직 분리
