+++ src/stores/mainStore.js
@@ -1,16 +1,20 @@
 import { create } from 'zustand';
 import { produce } from 'immer';
 import { getAllData, getData, updateData } from '@/api/CRUD';
-import { geocodeAddress, getUserLocation } from '@/utils';
+import {
+  geocodeAddress,
+  getUserLocation,
+  getDistanceFromLatLonInKm,
+  loadPostcodeScript,
+  extractDistrict,
+} from '@/utils';
 import axios from 'axios';
 import pb from '@/api/pb';
 import { useMapStore } from './mapStore';
-const MapUrl = import.meta.env.VITE_KAKAO_POSTCODE_SCRIPT_URL;
 
 export const mainStore = create((set) => {
   const INITIAL_STATE = {
     searchInput: {
-      searchWord: '',
       gymsList: [],
       filterGyms: [],
       isGymsLoaded: false,
@@ -18,42 +22,9 @@ export const mainStore = create((set) => {
       gymData: {},
       trainerList: {},
       trainerData: {},
-      updatedFilters: [],
       wishList: [],
       wishListChecked: {},
       filteredGymsByDistance: [],
-    },
-    searchFilter: {
-      rating: {
-        star1: false,
-        star2: false,
-        star3: false,
-        star4: false,
-        star5: false,
-      },
-      healthPrice: {
-        monthly10: false,
-        monthly15: false,
-        monthly20: false,
-      },
-      ptPrice: {
-        pt50: false,
-        pt60: false,
-        pt70: false,
-      },
-      amenities: {
-        parking: false,
-        showerRoom: false,
-        gxRoom: false,
-        wifi: false,
-        locker: false,
-        clothes: false,
-      },
-      trainerCount: {
-        oneToTwo: false,
-        threeToFour: false,
-        fiveToSix: false,
-      },
     },
     selectedLocation: '위치를 불러오는 중...',
     locationLoading: true,
@@ -67,243 +38,14 @@ export const mainStore = create((set) => {
     currentSwiperTrainerId: '',
   };
 
-  // 검색어 입력 처리
-  const handleSearchInput = (value) => {
+  const setFilteredGyms = (gyms) => {
     set(
       produce((draft) => {
-        draft.searchInput.searchWord = value;
-      })
-    );
-    const { filteredGymsByDistance } = mainStore.getState().searchInput;
-
-    if (!value.trim()) {
-      set(
-        produce((draft) => {
-          draft.searchInput.filterGyms = filteredGymsByDistance;
-        })
-      );
-      return;
-    }
-  };
-
-  // 검색 제출 처리
-  const handleSearchSubmit = (e) => {
-    e.preventDefault();
-
-    const { searchWord, filteredGymsByDistance } =
-      mainStore.getState().searchInput;
-
-    // 검색어가 없으면 필터 적용된 헬스장 리스트 반환
-    if (!searchWord.trim()) return;
-
-    const searchLower = searchWord.toLowerCase().trim();
-
-    // 검색어에서 '일일권', '원' 등 제거하고 숫자를 추출
-    const searchPrice = parseInt(
-      searchLower.replace(/[^0-9]/g, '') // 숫자 이외의 문자 제거
-    );
-
-    // 헬스장 이름, 주소, 일일권 가격이 검색어와 일치하는 헬스장 필터링
-    const filteredGyms = filteredGymsByDistance.filter((gym) => {
-      const gymName = gym.name.toLowerCase();
-      const gymAddress = gym.address.toLowerCase();
-      const gymOnedayPrice = gym.oneDayPrice; // 숫자 그대로 비교
-
-      // 이름, 주소 또는 일일권 가격과 검색어 비교
-      return (
-        gymName.includes(searchLower) ||
-        gymAddress.includes(searchLower) ||
-        (searchPrice && gymOnedayPrice === searchPrice)
-      );
-    });
-
-    // 필터링된 헬스장 리스트 업데이트
-    set(
-      produce((draft) => {
-        draft.searchInput.filterGyms = filteredGyms;
+        draft.searchInput.filterGyms = gyms;
       })
     );
   };
 
-  // 필터 버튼 체크 핸들러
-  const handleCheckboxChange = (target) => {
-    const { name } = target;
-    const filtername = target.getAttribute('data-filtername');
-
-    set(
-      produce((draft) => {
-        const isSelected = draft.searchFilter[filtername][name]; // 현재 선택된 상태 확인
-
-        if (filtername !== 'amenities') {
-          Object.keys(draft.searchFilter[filtername]).forEach((key) => {
-            draft.searchFilter[filtername][key] = false;
-          });
-        }
-
-        // 선택된 항목이 이미 true였다면 체크 해제, 아니면 선택
-        draft.searchFilter[filtername][name] = !isSelected;
-      })
-    );
-    updateCheckedFilters();
-  };
-
-  // 별점 필터를 처리하는 함수
-  const applyRatingFilter = (gym, filterNames) => {
-    const rating = gym.rating; // gym 객체에서 별점 값 가져오기
-    return filterNames.some((filterName) => {
-      const starRating = parseInt(filterName.replace('star', ''), 10);
-      return rating >= starRating; // 별점이 해당 필터보다 크거나 같은 경우
-    });
-  };
-  // 가격 필터를 처리하는 함수
-  const applyPriceFilter = (gym, filterNames) => {
-    const gymPrices = gym.healthPrice || {};
-
-    const oneMonthPrice = gymPrices['1Month']
-      ? parseInt(gymPrices['1Month'], 10)
-      : null;
-
-    if (!oneMonthPrice) return;
-
-    const filters = filterNames.every((filterName) => {
-      const maxPrice = parseInt(filterName.replace('monthly', ''), 10) * 10000;
-      return oneMonthPrice <= maxPrice;
-    });
-    return filters;
-  };
-  // pt가격 필터를 처리하는 함수
-  const applyPtPriceFilter = (gym, filterNames) => {
-    const ptPrice = gym.PtPrice || {};
-
-    const tenSessionsPrice = ptPrice['10Sessions']
-      ? parseInt(ptPrice['10Sessions'], 10)
-      : null;
-
-    if (!tenSessionsPrice) return;
-
-    const filter = filterNames.every((filterName) => {
-      const maxPrice = parseInt(filterName.replace('pt', ''), 10) * 10000;
-      return tenSessionsPrice <= maxPrice;
-    });
-
-    return filter;
-  };
-  // 편의시설 필터 처리 함수
-  const applyAmenitiesFilter = (gym, filterNames) => {
-    // 편의시설 데이터를 가져옵니다
-    const amenities = gym.amenities || {};
-    // 필터링된 편의시설 체크
-    return filterNames.every((filterName) => amenities[filterName] === true);
-  };
-
-  // 필터링된 헬스장 목록 업데이트 함수
-  const updateCheckedFilters = () => {
-    const { searchFilter } = mainStore.getState();
-    const { gymsList, filteredGymsByDistance } =
-      mainStore.getState().searchInput;
-    const checkedFilters = {};
-
-    // 체크된 필터 추출
-    Object.entries(searchFilter).forEach(([filterCategory, filters]) => {
-      Object.entries(filters).forEach(([filterName, isChecked]) => {
-        if (isChecked) {
-          if (!checkedFilters[filterCategory]) {
-            checkedFilters[filterCategory] = [];
-          }
-          checkedFilters[filterCategory].push(filterName);
-        }
-      });
-    });
-
-    // 필터링된 헬스장 목록 업데이트
-    const filteredGyms =
-      Object.keys(checkedFilters).length === 0
-        ? gymsList // 필터가 없으면 원래 리스트로
-        : filteredGymsByDistance.filter((gym) => {
-            return Object.entries(checkedFilters).every(
-              ([filterCategory, filterNames]) => {
-                if (filterCategory === 'rating') {
-                  // 별점 필터 처리
-                  return applyRatingFilter(gym, filterNames);
-                }
-                if (filterCategory === 'healthPrice') {
-                  // 가격 필터 처리
-                  return applyPriceFilter(gym, filterNames);
-                }
-                if (filterCategory === 'PtPrice') {
-                  // PT 가격 필터 처리
-                  return applyPtPriceFilter(gym, filterNames);
-                }
-                if (filterCategory === 'amenities') {
-                  // 편의시설 필터 처리
-                  return applyAmenitiesFilter(gym, filterNames);
-                }
-                // 다른 필터 카테고리 처리
-                return filterNames.some(
-                  (filterName) => gym[filterCategory] === filterName
-                );
-              }
-            );
-          });
-
-    // 상태 업데이트
-    set(
-      produce((draft) => {
-        draft.checkedFilters = checkedFilters;
-        draft.searchInput.filterGyms = filteredGyms;
-      })
-    );
-  };
-
-  const handleSelectedFilters = () => {
-    const amenitiesMapping = {
-      parking: '주차장',
-      wifi: 'Wi-Fi',
-      showerRoom: '샤워실',
-      locker: '개인락커',
-      clothes: '운동복',
-      gxRoom: 'GX룸',
-    };
-    set(
-      produce((draft) => {
-        // 기존에 저장된 updatedFilters를 초기화
-        draft.searchInput.updatedFilters = [];
-
-        // searchFilter를 순회하며 true인 필터 추출 및 변환
-        Object.keys(draft.searchFilter).forEach((key) => {
-          const trueFilters = Object.entries(draft.searchFilter[key])
-            .filter(([, value]) => value === true)
-            .map(([filterName]) => {
-              // 별점 필터 변환
-              if (filterName.startsWith('star')) {
-                const stars = filterName.replace('star', '');
-                return `별점 ${stars}점`;
-              }
-              // 헬스 가격 필터 변환
-              if (filterName.startsWith('monthly')) {
-                const price = filterName.replace('monthly', '');
-                return `월 ${price}만원`;
-              }
-              // PT 가격 필터 변환
-              if (filterName.startsWith('pt')) {
-                const price = filterName.replace('pt', '');
-                return `PT 10회 ${price}만원`;
-              }
-              // 어메니티 필터 변환 (편의시설)
-              if (amenitiesMapping[filterName]) {
-                return amenitiesMapping[filterName];
-              }
-              return filterName; // 변환 불가한 필터는 그대로 반환
-            });
-
-          // 변환된 필터가 있으면 updatedFilters 배열에 추가
-          if (trueFilters.length > 0) {
-            draft.searchInput.updatedFilters.push(...trueFilters);
-          }
-        });
-      })
-    );
-  };
   // getGymsList 함수에서 위치를 매개변수로 받도록 수정
   const getGymsList = async (latitude, longitude) => {
     try {
@@ -367,42 +109,6 @@ export const mainStore = create((set) => {
       })
     );
   };
-
-  const extractDistrict = (address) => {
-    // 주소를 공백을 기준으로 나눈 후, '구'나 '시'가 포함된 위치까지 추출
-    const addressParts = address.split(' ');
-
-    // '구' 또는 '시'로 끝나는 부분을 찾고 그 이전까지 추출
-    const endIndex = addressParts.findIndex(
-      (part) => part.includes('구') || part.includes('시')
-    );
-
-    // '구'가 포함된 부분까지 잘라내어 반환
-    if (endIndex !== -1) {
-      return addressParts.slice(0, endIndex + 1).join(' ');
-    }
-    // '구'가 없으면 전체 반환 (필요에 따라 동으로 처리 가능)
-    return address;
-  };
-
-  // Haversine 공식을 사용한 거리 계산 함수
-  const getDistanceFromLatLonInKm = (lat1, lon1, lat2, lon2) => {
-    const R = 6371; // 지구 반지름 (단위: km)
-    const dLat = deg2rad(lat2 - lat1);
-    const dLon = deg2rad(lon2 - lon1);
-    const a =
-      Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-      Math.cos(deg2rad(lat1)) *
-        Math.cos(deg2rad(lat2)) *
-        Math.sin(dLon / 2) *
-        Math.sin(dLon / 2);
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-    const distance = R * c; // 두 지점 사이 거리 (단위: km)
-    return distance;
-  };
-
-  // 각도를 라디안으로 변환하는 함수
-  const deg2rad = (deg) => deg * (Math.PI / 180);
 
   // 헬스장 리스트 필터링 및 정렬 함수
   const filterGymsByDistance = async (gyms, userLat, userLon) => {
@@ -471,28 +177,6 @@ export const mainStore = create((set) => {
         })
       );
     }
-  };
-
-  // 주소 검색 스크립트 로딩
-  const loadPostcodeScript = () => {
-    return new Promise((resolve, reject) => {
-      if (window.daum && window.daum.Postcode) {
-        resolve(); // 이미 스크립트가 로드된 경우
-        return;
-      }
-
-      const existingScript = document.querySelector(`script[src="${MapUrl}"]`);
-      if (existingScript) {
-        resolve();
-        return;
-      }
-
-      const script = document.createElement('script');
-      script.src = MapUrl;
-      script.onload = () => resolve();
-      script.onerror = () => reject(new Error('주소 검색 스크립트 로딩 실패'));
-      document.body.appendChild(script);
-    });
   };
 
   // 카카오 주소 검색 스크립트 URL
@@ -782,12 +466,7 @@ export const mainStore = create((set) => {
   return {
     ...INITIAL_STATE,
     handleMethod: {
-      handleCheckboxChange,
-      handleSearchInput,
-      handleSearchSubmit,
-      handleSelectedFilters,
       fetchGymDetails,
-      updateCheckedFilters,
       getGymsList,
       getCurrentLocation,
       searchLocation,
@@ -799,6 +478,7 @@ export const mainStore = create((set) => {
       setTrainerDetailPath,
       setSelectedTrainerId,
       handleTrainerSwiperChange,
+      setFilteredGyms,
     },
   };
 });
